Forget a type cast

+++ c++/jksn.cpp
@@ -1277,7 +1277,7 @@ static std::string UTF8ToUTF16LE(const std::string &utf8str, bool strict) {
             }
         } else if(uint8_t(utf8str[i]) < 0xf0) {
             if(UTF8CheckContinuation(utf8str, i, 2)) {
-                uint32_t ucs4 = uint32_t(utf8str[i] & 0xf) << 12 | uint32_t(utf8str[i+1] & 0x3f) << 6 | (utf8str[i+2] & 0x3f);
+                uint32_t ucs4 = uint32_t(utf8str[i] & 0xf) << 12 | uint32_t(utf8str[i+1] & 0x3f) << 6 | uint32_t(utf8str[i+2] & 0x3f);
                 if(ucs4 >= 0x800 && (ucs4 & 0xf800) != 0xd800) {
                     utf16str.append({
                         char(ucs4),
